requirements.txt
gunicorn
django-heroku
Django
djangorestframework
